% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tympo.gl}
\alias{tympo.gl}
\title{A genlight object created via /data-raw/data.R}
\format{
genlight object
}
\usage{
tympo.gl
}
\description{
This a test data set used for the dartR tutorial on data
It is a genlight object of the Canberra grassland earless dragon
}
\author{
Emily but originally provided by Bernd Gruber (bugs? Post to
\url{https://groups.google.com/d/forum/dartr}
}
\keyword{datasets}
